Scale and scroll now use the Double precision

+++ mapcompose/src/main/java/ovh/plrapps/mapcompose/api/DefaultCanvas.kt
@@ -29,15 +29,15 @@ fun DefaultCanvas(
     ) {
         withTransform({
             /* Geometric transformations seem to be applied in reversed order of declaration */
-            translate(left = -mapState.scroll.x, top = -mapState.scroll.y)
+            translate(left = -mapState.scroll.x.toFloat(), top = -mapState.scroll.y.toFloat())
             rotate(
                 degrees = mapState.rotation,
                 pivot = Offset(
-                    x = mapState.centroidX.toFloat() * mapState.fullSize.width * mapState.scale,
-                    y = mapState.centroidY.toFloat() * mapState.fullSize.height * mapState.scale
+                    x = (mapState.centroidX * mapState.fullSize.width * mapState.scale).toFloat(),
+                    y = (mapState.centroidY.toFloat() * mapState.fullSize.height * mapState.scale).toFloat()
                 )
             )
-            scale(scale = mapState.scale, Offset.Zero)
+            scale(scale = mapState.scale.toFloat(), Offset.Zero)
         }, drawBlock)
     }
 }
